requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
